Evaluate signal tracks in native runtime
