+++ crates/noon-runtime/src/reactive_impl.rs
@@ -1,0 +1,421 @@
+use std::collections::BTreeMap;
+
+use noon_compile::{CompileError, CompiledScene};
+use noon_core::{
+    ObjectId, Property, ReactiveBinding, ReactiveError, ReactiveEvaluationStats, ReactiveProgram,
+    ReactiveState, ReactiveValue, SemanticScene, SignalId,
+};
+
+use crate::{FrameState, SceneInstance};
+
+#[derive(Clone, Debug, PartialEq)]
+pub enum SceneBuildError {
+    Compile(CompileError),
+    Reactive(ReactiveError),
+}
+
+impl std::fmt::Display for SceneBuildError {
+    fn fmt(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::Compile(error) => write!(formatter, "scene compilation failed: {error}"),
+            Self::Reactive(error) => write!(formatter, "reactive compilation failed: {error}"),
+        }
+    }
+}
+
+impl std::error::Error for SceneBuildError {}
+
+impl From<CompileError> for SceneBuildError {
+    fn from(value: CompileError) -> Self {
+        Self::Compile(value)
+    }
+}
+
+impl From<ReactiveError> for SceneBuildError {
+    fn from(value: ReactiveError) -> Self {
+        Self::Reactive(value)
+    }
+}
+
+/// Work performed by the most recent native reactive input update.
+///
+/// These counters deliberately exclude total scene/object counts. Once a semantic
+/// scene is lowered, an input update performs dependency evaluation plus direct
+/// writes to precomputed dense frame targets only.
+#[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
+pub struct ReactiveRuntimeStats {
+    pub derived_signals_evaluated: usize,
+    pub bindings_invalidated: usize,
+    pub dense_targets_applied: usize,
+    pub dense_targets_changed: usize,
+}
+
+#[derive(Clone, Copy, Debug)]
+struct ReactiveTarget {
+    signal: SignalId,
+    object_index: usize,
+    property: Property,
+}
+
+#[derive(Clone, Debug)]
+pub(crate) struct ReactiveRuntime {
+    state: ReactiveState,
+    targets: Vec<ReactiveTarget>,
+    target_lookup: BTreeMap<(u64, u8), usize>,
+    targets_by_object: Vec<Vec<usize>>,
+}
+
+impl ReactiveRuntime {
+    fn new(
+        compiled: &CompiledScene,
+        bindings: &[ReactiveBinding],
+        program: ReactiveProgram,
+    ) -> Self {
+        let mut targets = Vec::with_capacity(bindings.len());
+        let mut target_lookup = BTreeMap::new();
+        let mut targets_by_object = vec![Vec::new(); compiled.objects().len()];
+
+        for binding in bindings {
+            let object_index = compiled
+                .object_index(binding.object)
+                .expect("reactive graph was validated against the compiled definition")
+                as usize;
+            let target_index = targets.len();
+            targets.push(ReactiveTarget {
+                signal: binding.signal,
+                object_index,
+                property: binding.property,
+            });
+            target_lookup.insert(binding_key(binding.object, binding.property), target_index);
+            targets_by_object[object_index].push(target_index);
+        }
+
+        Self {
+            state: program.instantiate(),
+            targets,
+            target_lookup,
+            targets_by_object,
+        }
+    }
+
+    fn target(&self, object: ObjectId, property: Property) -> ReactiveTarget {
+        let index = *self
+            .target_lookup
+            .get(&binding_key(object, property))
+            .expect("reactive update must reference a lowered binding");
+        self.targets[index]
+    }
+}
+
+impl SceneInstance {
+    /// Compile a high-level semantic scene once and attach its validated native
+    /// reactive program to this runtime instance.
+    ///
+    /// Reactive bindings are lowered to dense frame object indices here. Later
+    /// input updates therefore do not rebuild `SceneDefinition`, recompile the
+    /// timeline, or scan unrelated objects.
+    pub fn from_semantic(scene: &SemanticScene) -> Result<Self, SceneBuildError> {
+        let compiled = CompiledScene::compile(scene.definition())?;
+        let program = scene.compile_reactive()?;
+        let reactive = ReactiveRuntime::new(&compiled, scene.reactive().bindings(), program);
+        let mut instance = Self::new(compiled);
+        instance.reactive = Some(reactive);
+        instance.reapply_reactive();
+        Ok(instance)
+    }
+
+    pub const fn last_reactive_stats(&self) -> ReactiveRuntimeStats {
+        self.last_reactive_stats
+    }
+
+    pub fn reactive_value(&self, signal: SignalId) -> Option<&ReactiveValue> {
+        self.reactive.as_ref()?.state.value(signal)
+    }
+
+    /// Change one native input and apply only its invalidated bindings to the
+    /// already-compiled dense frame state.
+    pub fn set_reactive_input(
+        &mut self,
+        signal: SignalId,
+        value: impl Into<ReactiveValue>,
+    ) -> Result<&FrameState, ReactiveError> {
+        let update = self
+            .reactive
+            .as_mut()
+            .ok_or(ReactiveError::UnknownSignal(signal))?
+            .state
+            .set_input(signal, value)?;
+        let evaluation = update.stats();
+        let mut changed_targets = 0;
+
+        for change in update.property_changes() {
+            let target = self
+                .reactive
+                .as_ref()
+                .expect("reactive state exists while applying its update")
+                .target(change.object, change.property);
+            if apply_reactive_value(
+                &mut self.frame,
+                target.object_index,
+                target.property,
+                &change.value,
+            ) {
+                self.changes.insert(target.object_index);
+                changed_targets += 1;
+            }
+        }
+
+        self.last_reactive_stats =
+            runtime_stats(evaluation, update.property_changes().len(), changed_targets);
+        Ok(&self.frame)
+    }
+
+    pub(crate) fn reapply_reactive(&mut self) {
+        let Self {
+            reactive, frame, ..
+        } = self;
+        let Some(reactive) = reactive.as_ref() else {
+            return;
+        };
+        for target in &reactive.targets {
+            let value = reactive
+                .state
+                .value(target.signal)
+                .expect("lowered reactive target references a valid signal");
+            apply_reactive_value(frame, target.object_index, target.property, value);
+        }
+    }
+
+    pub(crate) fn reapply_reactive_for_object(&mut self, object_index: usize) {
+        let Self {
+            reactive, frame, ..
+        } = self;
+        let Some(reactive) = reactive.as_ref() else {
+            return;
+        };
+        let Some(target_indices) = reactive.targets_by_object.get(object_index) else {
+            return;
+        };
+        for target_index in target_indices {
+            let target = reactive.targets[*target_index];
+            let value = reactive
+                .state
+                .value(target.signal)
+                .expect("lowered reactive target references a valid signal");
+            apply_reactive_value(frame, target.object_index, target.property, value);
+        }
+    }
+}
+
+fn runtime_stats(
+    evaluation: ReactiveEvaluationStats,
+    dense_targets_applied: usize,
+    dense_targets_changed: usize,
+) -> ReactiveRuntimeStats {
+    ReactiveRuntimeStats {
+        derived_signals_evaluated: evaluation.derived_signals_evaluated,
+        bindings_invalidated: evaluation.bindings_invalidated,
+        dense_targets_applied,
+        dense_targets_changed,
+    }
+}
+
+fn binding_key(object: ObjectId, property: Property) -> (u64, u8) {
+    (object.get(), property_slot(property))
+}
+
+const fn property_slot(property: Property) -> u8 {
+    match property {
+        Property::Presence => 0,
+        Property::Transform => 1,
+        Property::Position => 2,
+        Property::Rotation => 3,
+        Property::Opacity => 4,
+        Property::Appearance => 5,
+        Property::Reveal => 6,
+        Property::Morph => 7,
+    }
+}
+
+fn apply_reactive_value(
+    frame: &mut FrameState,
+    object_index: usize,
+    property: Property,
+    value: &ReactiveValue,
+) -> bool {
+    match (property, value) {
+        (Property::Presence, ReactiveValue::Bool(value)) => {
+            let changed = frame.presences[object_index] != *value;
+            frame.presences[object_index] = *value;
+            changed
+        }
+        (Property::Position, ReactiveValue::Vec2(value)) => {
+            let object = &mut frame.objects[object_index];
+            let changed = object.transform.translation != *value;
+            object.transform.translation = *value;
+            changed
+        }
+        (Property::Rotation, ReactiveValue::Scalar(value)) => {
+            let object = &mut frame.objects[object_index];
+            let changed = object.transform.rotation != *value;
+            object.transform.rotation = *value;
+            changed
+        }
+        (Property::Opacity, ReactiveValue::Scalar(value)) => {
+            let object = &mut frame.objects[object_index];
+            let changed = object.style.opacity != *value;
+            object.style.opacity = *value;
+            changed
+        }
+        (Property::Appearance, ReactiveValue::Scalar(value)) => {
+            let value = value.clamp(0.0, 1.0);
+            let object = &mut frame.objects[object_index];
+            let changed = object.appearance != value;
+            object.appearance = value;
+            changed
+        }
+        (Property::Reveal, ReactiveValue::Scalar(value)) => {
+            let value = value.clamp(0.0, 1.0);
+            let changed = frame.reveals[object_index] != value;
+            frame.reveals[object_index] = value;
+            changed
+        }
+        (Property::Morph, ReactiveValue::Scalar(value)) => {
+            let value = value.clamp(0.0, 1.0);
+            let changed = frame.morphs[object_index] != value;
+            frame.morphs[object_index] = value;
+            changed
+        }
+        (Property::Transform, _) => {
+            unreachable!("reactive values cannot drive object-snapshot Transform properties")
+        }
+        _ => unreachable!("validated reactive binding value type must match its property"),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use noon_core::{GeometryRef, RateFunction, ReactiveExpr, TrackTiming, Vec2};
+
+    use super::*;
+
+    #[test]
+    fn initial_reactive_values_are_lowered_into_frame_state() {
+        let mut scene = SemanticScene::new();
+        let object = scene.add(GeometryRef::circle(1.0));
+        let position = scene.add_input(Vec2::new(3.0, -2.0));
+        scene.bind(position, object, Property::Position);
+
+        let instance = SceneInstance::from_semantic(&scene).expect("semantic scene must compile");
+        assert_eq!(
+            instance.frame().objects[0].transform.translation,
+            Vec2::new(3.0, -2.0)
+        );
+    }
+
+    #[test]
+    fn input_update_mutates_only_lowered_dense_target() {
+        let mut scene = SemanticScene::new();
+        let untouched = scene.add(GeometryRef::circle(1.0));
+        let target = scene.add(GeometryRef::circle(1.0));
+        let input = scene.add_input(1.0_f32);
+        let doubled = scene.add_derived(ReactiveExpr::Mul(
+            Box::new(ReactiveExpr::signal(input)),
+            Box::new(ReactiveExpr::scalar(2.0)),
+        ));
+        scene.bind(doubled, target, Property::Rotation);
+
+        let mut instance =
+            SceneInstance::from_semantic(&scene).expect("semantic scene must compile");
+        instance.take_frame_changes();
+        instance
+            .set_reactive_input(input, 2.0_f32)
+            .expect("input update must work");
+
+        assert_eq!(instance.frame().objects[0].id, untouched);
+        assert_eq!(instance.frame().objects[0].transform.rotation, 0.0);
+        assert_eq!(instance.frame().objects[1].id, target);
+        assert_eq!(instance.frame().objects[1].transform.rotation, 4.0);
+        assert_eq!(instance.take_frame_changes().object_indices(), &[1]);
+        assert_eq!(
+            instance.last_reactive_stats(),
+            ReactiveRuntimeStats {
+                derived_signals_evaluated: 1,
+                bindings_invalidated: 1,
+                dense_targets_applied: 1,
+                dense_targets_changed: 1,
+            }
+        );
+    }
+
+    #[test]
+    fn seeks_reapply_reactive_values_after_timeline_evaluation() {
+        let mut scene = SemanticScene::new();
+        let object = scene.add(GeometryRef::circle(1.0));
+        scene
+            .definition_mut()
+            .animate_position(
+                object,
+                Vec2::ZERO,
+                Vec2::new(10.0, 0.0),
+                TrackTiming::new(0.0, 2.0, RateFunction::Linear),
+            )
+            .expect("timeline must be valid");
+        let rotation = scene.add_input(0.25_f32);
+        scene.bind(rotation, object, Property::Rotation);
+
+        let mut instance =
+            SceneInstance::from_semantic(&scene).expect("semantic scene must compile");
+        instance
+            .set_reactive_input(rotation, 1.25_f32)
+            .expect("input update must work");
+        instance.seek(1.0).expect("seek must work");
+
+        assert_eq!(
+            instance.frame().objects[0].transform.translation,
+            Vec2::new(5.0, 0.0)
+        );
+        assert_eq!(instance.frame().objects[0].transform.rotation, 1.25);
+    }
+
+    #[test]
+    fn reactive_update_cost_does_not_scale_with_static_object_count() {
+        let mut scene = SemanticScene::new();
+        for _ in 0..50_000 {
+            scene.add(GeometryRef::circle(1.0));
+        }
+        let target = scene.add(GeometryRef::circle(1.0));
+        let input = scene.add_input(1.0_f32);
+        let doubled = scene.add_derived(ReactiveExpr::Mul(
+            Box::new(ReactiveExpr::signal(input)),
+            Box::new(ReactiveExpr::scalar(2.0)),
+        ));
+        let shifted = scene.add_derived(ReactiveExpr::Add(
+            Box::new(ReactiveExpr::signal(doubled)),
+            Box::new(ReactiveExpr::scalar(1.0)),
+        ));
+        scene.bind(shifted, target, Property::Rotation);
+
+        let mut instance =
+            SceneInstance::from_semantic(&scene).expect("semantic scene must compile");
+        instance.take_frame_changes();
+        let timeline_stats_before = instance.last_stats();
+        instance
+            .set_reactive_input(input, 3.0_f32)
+            .expect("input update must work");
+
+        assert_eq!(instance.frame().objects[50_000].id, target);
+        assert_eq!(instance.frame().objects[50_000].transform.rotation, 7.0);
+        assert_eq!(instance.take_frame_changes().object_indices(), &[50_000]);
+        assert_eq!(instance.last_stats(), timeline_stats_before);
+        assert_eq!(
+            instance.last_reactive_stats(),
+            ReactiveRuntimeStats {
+                derived_signals_evaluated: 2,
+                bindings_invalidated: 1,
+                dense_targets_applied: 1,
+                dense_targets_changed: 1,
+            }
+        );
+    }
+}
